fixed issues in the rebase

+++ server/model/kino.go
@@ -8,7 +8,7 @@ import (
 
 // Kino stores all movies
 type Kino struct {
-	Id          int32       `gorm:"primary_key;AUTO_INCREMENT;column:kino;type:int;not null;"`
+	Id          int64       `gorm:"primary_key;AUTO_INCREMENT;column:kino;type:int;not null;"`
 	Date        time.Time   `gorm:"column:date;type:datetime;not null;"`
 	Created     time.Time   `gorm:"column:created;type:timestamp;not null;default:CURRENT_TIMESTAMP"`
 	Title       string      `gorm:"column:title;type:text;not null;"`
@@ -20,7 +20,7 @@ type Kino struct {
 	ImdbRating  string      `gorm:"column:rating;type:varchar(4);not null;"`
 	Description string      `gorm:"column:description;type:text;not null;"`
 	Trailer     null.String `gorm:"column:trailer"`
-	FilesID     int32       `gorm:"column:cover"`
+	FilesID     int64       `gorm:"column:cover"`
 	Files       Files       `gorm:"foreignKey:FilesID;references:file"`
 	Link        string      `gorm:"column:link;type:varchar(190);not null;unique;"`
 }
